Add a basic query parser to handle single word dismax queries
Dismax handler adds PF boosts to queries with a single word, which the edismax parser does not.
This creates a difference in search results.
To fix this, we add a basic handler to handle single word dismax queries. The perl layer in MBS will provide the proper abstractions against the dismax query parser for single and multi word queries.