Use the new MusicBrainz tokenizers and filters
The previous analyzer pipeline, although it preserved the original token, created exponential sub queries, since it basically copied each token because of the preserveOriginal flag. This lead to Solr getting under heavy load.

Using this new tokenizer/filer, which are basically an almost replica of the old search server ones, except for the fact that the new MBSFilter replaces ALHPANUMPUNCT's non alpha chars with spaces so that trim filter can work on it instead of Word Delim Filter which has been deprecated.

This allows for us to search for purely punctuational tokens like "!!!" while still filtering something like "as!!" to "as"