Add a default sort to all cores to break equal score tiebreakers
MBS-12154
In some kinds of searches (such as "release-group arid:[mbid]"), the score
of all resulting documents is constant. If there are more documents
matching this result than the page size, it becomes difficult to
consistently page through the results.
To fix this, add an arbitrary sort after the score to break it in case
that it is equal. I just chose the mbid or id for each core, as it's
also the primary key and is guaranteed to be in the document and set.